Add sub-blocks.
This is another level of tiling between blocks and vector-clusters.
This allows threads in a team to work on arbitrarily-sized tiles.
Previously, they could only work on slabs or pencils within a block.
Grouping for sub-blocks is supported with cmd-line options to control group size.

Also: remove some g++ warnings.