Enable default folding, including compiler API to set FP precision.
Folding is now computed by the stencil compiler if not specified.
If folding is provided but wrong, it is adjusted.
This required the compiler to know the size of each FP element.
Added -elem-bytes option for this.
Printer classes report SIMD size for each print format.
Updated compiler API docs.

Also,
- Changed -st and -r compiler options to -stencil and -radius to be consistent with other utilities.
- Moved src/foldBuilder/Tuple.hpp to common src/tuple.hpp to be used by grids.
- Refactored tuple.hpp to add Scalar<T> template to be used for one named value.
- Scalar objects now used for directions and single indices.
- Required changing ExprTuple to ExprScalar.
- This is just initial part of grid refactoring to allow grids to be created via API.